+++ src/precision.h
@@ -1,0 +1,2 @@
+// typedef long double ldouble;
+typedef double ldouble;