+++ tests/testthat/test_MC_WUC.R
@@ -2,10 +2,14 @@
 #
 # Hua Zhong
 # Created: Aug 11, 2019
+# Modified:
+#   March 13, 2020. MS. Changed the random data from uniform integer
+#     to standard normal to avoid examples with multile optimal
+#     solutions occuring to integers.
 
 library(testthat)
 library(Ckmeans.1d.dp)
-context("Ckmeans.1d.dp")
+context("MDW_Ckmeans")
 
 test_that("Test 1D weights comparing Ckmeans and MDW_Ckmeans", {
   total <- 10000
@@ -13,12 +17,15 @@ test_that("Test 1D weights comparing Ckmeans and MDW_Ckmeans", {
   i <- 0
   for(j in c(1:total)){
 
-    X <- sort(sample(x = c(1:100), size = 20, replace = TRUE))
-    Y <- matrix(sample(x = c(1:100), size = 20, replace = TRUE), ncol=1, nrow=length(X))
+    # X <- sort(sample(x = c(1:100), size = 20, replace = TRUE)) # MS March 13, 2020
+    X <- sort(sample(x = rnorm(100), size = 20, replace = TRUE)) # MS March 13, 2020
+
+    Y <- matrix(sample(x = c(1:100), size = 20, replace = TRUE),
+                ncol=1, nrow=length(X))
 
     cluster.num <- c(5:10)
 
-    res.1 <- MultiChannel.WUC(x = X, y = Y, k =cluster.num)
+    res.1 <- MultiChannel.WUC(x = X, y = Y, k = cluster.num)
 
     res.2 <- suppressWarnings(
       Ckmeans.1d.dp(x = X, y = as.numeric(Y), k = cluster.num,
@@ -38,7 +45,7 @@ test_that("Test 1D weights comparing Ckmeans and MDW_Ckmeans", {
       ))) {
       i <- i + 1
     }else{
-      #stop()
+      # stop()
     }
   }
   # print(i)